refactor: move src/cacheCreate.js to src/cache/lru.js
